requirements.txt
psycopg2==2.9.3
cassandra-driver==3.25.0
python-dotenv==0.20.0
